requirements.txt
colorama==0.4.3
gitpython==3.1.27
pygls==0.13.0
doxysphinx==3.2.1
Sphinx==5.3.0
sphinx-data-viewer==0.1.2
sphinx-needs==1.1.0
sphinx-rtd-theme==1.1.1
sphinx-test-reports==1.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-doxylink==1.12.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-plantuml==0.24.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5